+++ addOns/authhelper/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [0.11.0] - 2024-01-10
 ### Changed
 - Maintenance changes.
 - Dropped "to Clipboard" from ZAP copy menu items or buttons (Issue 8179).
@@ -77,6 +77,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Support of authentication request identification and configuration.
 
+[0.11.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.11.0
 [0.10.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.10.0
 [0.9.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.9.0
 [0.8.0]: https://github.com/zaproxy/zap-extensions/releases/authhelper-v0.8.0